move find_models into common.py
